+++ src/mock/mock-destination.js
@@ -1,5 +1,5 @@
-import { getRandomValue } from '../utils.js';
-import { CITIES, DESCRIPTION } from './mock-const.js';
+import { getRandomValue } from '../utils/common.js';
+import { CITIES, DESCRIPTION } from './const.js';
 
 
 function generateDestination () {
