+++ src/mock/utils.js
@@ -1,6 +1,6 @@
 import dayjs from 'dayjs';
-import { getRandomInteger } from '../utils.js';
-import { Duration } from './mock-const.js';
+import { getRandomInteger } from '../utils/common.js';
+import { Duration } from './const.js';
 
 
 let date = dayjs().subtract(getRandomInteger(0, Duration.DAY), 'day').toDate();
